Ackley fixed. Benchmark functions updated
